Admin messages over

+++ resources/assets/js/components/Admin/Messages.vue
@@ -4,7 +4,6 @@
             <div class="head">
                 <div class="subject">{{ message.subject }}</div>
                 <i class="fa fa-close" @click="deleteMessage(index,message.id)"></i>
-                <i class="fa fa-reply"></i>
             </div>
             <div class="body">{{ message.message }}</div>
             <div class="email">{{ message.email }}</div>
@@ -32,10 +31,33 @@
 
         methods: {
             deleteMessage(index,message_id){
-                axios.delete('/admin/messages',
-                    {params: {'id': message_id}}
-                );
-                this.contactItems.splice(index,1);
+                var self = this;
+                this.$swal({
+                    title: 'Emin misin?',
+                    text: "Sildiğin mesajı geri alamazsın!",
+                    type: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Evet, sil!',
+                    cancelButtonText: 'Hayır'
+                }).then((result) => {
+                    if (result.value) {
+                        axios.delete('/admin/messages',{
+                            params: {id: message_id}
+                        }).then(response => {
+                            if(response.status === 200){
+                                self.$swal(
+                                    'Silindi!',
+                                    'Mesaj silindi.',
+                                    'success'
+                                );
+                                self.contactItems.splice(index,1);
+                            }
+                        });
+
+                    }
+                });
             }
         }
 
@@ -43,7 +65,7 @@
 </script>
 
 <style>
-    ul.messages i.fa.fa-close,ul.messages i.fa.fa-reply {
+    ul.messages i.fa.fa-close {
         cursor: pointer;
     }
 </style>
